+++ industrypool/providers.py
@@ -6,10 +6,10 @@ esi = ESIClientProvider(
     compatibility_date="2024-01-01",
     ua_appname="aa-industrypool",
     ua_version=__version__,
-    ua_url="https://github.com/your-org/aa-industrypool",
+    ua_url="https://github.com/evecarboot/aa-industrypool",
     operations=[
         "GetCorporationsCorporationIdDivisions",
         "GetCorporationsCorporationIdIndustryJobs",
-        "GetCorporationsCorporationIdAssetsBlueprints",
+        "GetCorporationsCorporationIdBlueprints",
     ],
 )