+++ test_urls.py
@@ -1,0 +1,5 @@
+"""Root URLconf for the test suite - industrypool is added by its Alliance Auth url hook."""
+
+from django.urls import include, path
+
+urlpatterns = [path("", include("allianceauth.urls"))]